#14 Introducing the ZugferdDocumentPdfMerger
